单独出来了应用相关的模块，便于第三方公司整合自己的应用中心
