+++ bistoury-application/pom.xml
@@ -1,0 +1,39 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+	<parent>
+		<artifactId>bistoury</artifactId>
+		<groupId>qunar.tc.bistoury</groupId>
+		<version>1.4.0-SNAPSHOT</version>
+	</parent>
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>bistoury-application</artifactId>
+	<packaging>pom</packaging>
+
+	<properties>
+		<skip_maven_deploy>false</skip_maven_deploy>
+	</properties>
+
+	<modules>
+		<module>bistoury-application-api</module>
+		<module>bistoury-application-mysql</module>
+	</modules>
+
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>qunar.tc.bistoury</groupId>
+				<artifactId>bistoury-application-api</artifactId>
+				<version>${project.parent.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>qunar.tc.bistoury</groupId>
+				<artifactId>bistoury-application-mysql</artifactId>
+				<version>${project.parent.version}</version>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
+
+</project>
