+++ bistoury-serverside-common/pom.xml
@@ -19,6 +19,10 @@
     <dependencies>
         <dependency>
             <groupId>qunar.tc.bistoury</groupId>
+            <artifactId>bistoury-application-api</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>qunar.tc.bistoury</groupId>
             <artifactId>bistoury-remoting</artifactId>
             <exclusions>
                 <exclusion>
