+++ bistoury-ui-service/pom.xml
@@ -21,6 +21,10 @@
             <artifactId>bistoury-common</artifactId>
         </dependency>
         <dependency>
+            <groupId>qunar.tc.bistoury</groupId>
+            <artifactId>bistoury-application-api</artifactId>
+        </dependency>
+        <dependency>
             <groupId>javax.servlet</groupId>
             <artifactId>javax.servlet-api</artifactId>
             <version>3.1.0</version>
