+++ bistoury-ui/src/main/java/qunar/tc/bistoury/ui/controller/AppController.java
@@ -22,13 +22,15 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+
+import qunar.tc.bistoury.application.api.AdminAppService;
+import qunar.tc.bistoury.application.api.AppServerService;
+import qunar.tc.bistoury.application.api.AppService;
 import qunar.tc.bistoury.serverside.bean.ApiResult;
 import qunar.tc.bistoury.serverside.bean.ApiStatus;
 import qunar.tc.bistoury.serverside.util.ResultHelper;
 import qunar.tc.bistoury.ui.security.LoginContext;
-import qunar.tc.bistoury.ui.service.AdminAppService;
-import qunar.tc.bistoury.ui.service.AppServerService;
-import qunar.tc.bistoury.ui.service.AppService;
+import qunar.tc.bistoury.ui.service.UserService;
 
 @Controller
 public class AppController {
@@ -40,6 +42,9 @@ public class AppController {
 
     @Autowired
     private AdminAppService adminAppService;
+
+    @Autowired
+    private UserService userService;
 
     @Autowired
     private AppServerService appServerService;
@@ -59,14 +64,14 @@ public class AppController {
     @ResponseBody
     public ApiResult isAdmin() {
         String userName = LoginContext.getLoginContext().getLoginUser();
-        return ResultHelper.success(adminAppService.isAdminUser(userName));
+        return ResultHelper.success(userService.isAdmin(userName));
     }
 
     @RequestMapping("searchApps")
     @ResponseBody
     public ApiResult searchApps(String searchAppKey) {
         String userName = LoginContext.getLoginContext().getLoginUser();
-        if (adminAppService.isAdminUser(userName)) {
+        if (userService.isAdmin(userName)) {
             return ResultHelper.success(adminAppService.searchApps(searchAppKey, ADMIN_PAGE_SIZE));
         } else {
             return ResultHelper.fail(ApiStatus.PERMISSION_DENY.getCode(), ApiStatus.PERMISSION_DENY.getMsg());
